09-09 Class 최종성적, 등급비율차트 기능 구현 + 좌측 메뉴 아이콘 수정, 클래스 토글기능

+++ src/main/java/com/example/app/model/dto/response/repository/TaskItemMapping.java
@@ -6,12 +6,13 @@ import com.example.app.model.domain.section.TaskItemInfo;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 public interface TaskItemMapping {
 
     Long getTaskItemIdx();
 
-    Long getTaskScore();
+    BigDecimal getTaskScore();
 
     Long getStudentStudentIdx();
